refactor: removed unnecessary worker
